Reorganize project structure and add contribution guidelines (#2)
* refactor: Reorganize project structure and add contribution guidelines

This commit reorganizes the project structure to formalize the development workflow. It introduces a 'development' directory to house the source code and moves the game scenes into a 'components' subdirectory as requested. The file paths in 'index.html' and 'main.js' have been updated to reflect this new structure.

Additionally, a CONTRIBUTING.md file has been created to document the new branching strategy and development workflow.

* refactor: Reorganize project structure and add contribution guidelines

This commit reorganizes the project structure to formalize the development workflow. It introduces a 'development' directory to house the source code and moves the game scenes into a 'components' subdirectory as requested. The file paths in 'index.html' and 'main.js' have been updated to reflect this new structure.

Additionally, a CONTRIBUTING.md file has been created to document the new branching strategy and development workflow.

---------
