+++ README.md
@@ -15,22 +15,42 @@ A web-based auto side-scrolling platformer game. Players will navigate through a
 3.  **Game Over:** If you hit an obstacle, the game will restart.
 
 ## Development
-This section is for developers who want to contribute to the project.
+
+This project is organized into two main directories:
+- `development/`: Contains the source code for the game. All development should happen here.
+- `production/`: Contains the built version of the game, ready for deployment. This directory is generated by the build script.
 
 ### Getting Started
-To run the game on your local machine, follow these steps:
+
+To run the game on your local machine for development, follow these steps:
 
 1.  Clone the repository to your local machine.
 2.  Navigate to the project directory in your terminal.
-3.  If this is your first time, make the test script executable:
+3.  Make the scripts executable (only needs to be done once):
     ```bash
-    chmod +x scripts/test.sh
+    chmod +x scripts/*.sh
     ```
-4.  Run the test script to start the local web server:
+4.  To run the development server:
     ```bash
-    ./scripts/test.sh
+    ./scripts/develop.sh
     ```
-5.  Open your web browser and navigate to `http://localhost:8000`.
+5.  Open your web browser and navigate to `http://localhost:8000`. Changes you make in the `development` directory will be reflected here.
+
+### Building for Production
+
+To create a production build, run the build script:
+```bash
+./scripts/build.sh
+```
+This will populate the `production` directory with the latest code from `development`.
+
+### Running in Production Mode
+
+To simulate running the game in a production environment, you can use the start script:
+```bash
+./scripts/start.sh
+```
+This serves the files from the `production` directory.
 
 ### Technologies
 This project will use the following technologies:
