+++ development/components/BootScene.js
@@ -1,0 +1,10 @@
+export default class BootScene extends Phaser.Scene {
+    constructor() {
+        super({ key: 'BootScene' });
+    }
+
+    create() {
+        console.log('BootScene is active. Starting PlayScene...');
+        this.scene.start('PlayScene');
+    }
+}
