+++ development/index.html
@@ -14,6 +14,6 @@
 </head>
 <body>
     <div id="game-container"></div>
-    <script src="js/game.js"></script>
+    <script type="module" src="main.js"></script>
 </body>
 </html>
